backend/requirements.txt
setuptools==59.5.0
numpy==1.26.4
openai-whisper==20240930
pyannote.audio==3.3.0
huggingface_hub==0.13.4
soundfile==0.13.1
speechbrain==0.5.16
python-dotenv==1.0.1
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
numba==0.59.1
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
asyncpg==0.29.0
pydub==0.25.1
reportlab==4.2.5
svglib==1.5.1